refactor watch-sync feature
