+++ src/app.js
@@ -43,13 +43,11 @@ import {
 } from './lib/ui/pip.js';
 
 import {
-  isYouTubeUrl,
-  getYouTubeId,
-  showYouTubePlayer,
-  hideYouTubePlayer,
-  getYTPlayer,
-  getYTReady,
-} from './lib/video/youtube.js';
+  toggleWatchMode as toggleWatchModeSync,
+  loadStream as loadStreamSync,
+  setupWatchSync,
+  getWatchMode,
+} from './lib/watch-sync.js';
 
 import { hasFrontAndBackCameras, switchCamera } from './lib/devices.js';
 
@@ -71,8 +69,6 @@ let isInitiator = false;
 let isAudioMuted = false;
 let isVideoOn = true;
 let currentFacingMode = 'user'; // 'user' = front, 'environment' = back
-let watchMode = false;
-let isSyncing = false; // Prevent sync loops
 
 function saveCurrentState() {
   saveState({
@@ -81,7 +77,7 @@ function saveCurrentState() {
     isAudioMuted,
     isVideoOn,
     currentFacingMode,
-    watchMode,
+    watchMode: getWatchMode(),
     wasConnected,
     streamUrl: streamUrlInput.value,
   });
@@ -176,12 +172,11 @@ function restoreUIState(savedState) {
   isVideoOn = savedState.isVideoOn;
   currentFacingMode = savedState.currentFacingMode;
   isInitiator = savedState.isInitiator;
-  watchMode = savedState.watchMode || false;
   wasConnected = savedState.wasConnected || false;
 
   if (!isVideoOn) toggleVideo();
   if (isAudioMuted) toggleMuteSelfMic();
-  if (savedState.watchMode && !watchMode) {
+  if (savedState.watchMode && !getWatchMode()) {
     toggleWatchMode();
     if (savedState.streamUrl) streamUrlInput.value = savedState.streamUrl;
   }
@@ -266,7 +261,7 @@ async function initiateChatRoom() {
   startChatBtn.disabled = true;
   hangUpBtn.disabled = false;
 
-  setupWatchSync();
+  setupWatchSync({ roomId, sharedVideo, streamUrlInput, syncStatus });
   toggleModeBtn.style.display = 'block';
 
   const role = 'initiator';
@@ -321,7 +316,7 @@ async function joinChatRoom(roomId) {
     peerConnection.addIceCandidate(new RTCIceCandidate(candidate));
   });
 
-  setupWatchSync();
+  setupWatchSync({ roomId, sharedVideo, streamUrlInput, syncStatus });
   toggleModeBtn.style.display = 'block';
   hangUpBtn.disabled = false;
 
@@ -369,7 +364,6 @@ async function hangUp() {
   sharedVideo.src = '';
   streamUrlInput.value = '';
   syncStatus.textContent = '';
-  watchMode = false;
   isAudioMuted = false;
   isVideoOn = true;
   wasConnected = false;
@@ -429,193 +423,21 @@ function toggleVideo() {
 }
 
 function toggleWatchMode() {
-  watchMode = ui.toggleWatchMode(
-    watchMode,
+  toggleWatchModeSync({
     videoContainer,
     watchContainer,
     toggleModeBtn,
     sharedVideo,
     streamUrlInput,
-    syncStatus
-  );
+    syncStatus,
+  });
   saveCurrentState();
 }
 
 function loadStream() {
   const url = streamUrlInput.value.trim();
-  if (!url) {
-    syncStatus.textContent = 'Please enter a stream URL';
-    return;
-  }
-  if (isYouTubeUrl(url)) {
-    const vid = getYouTubeId(url);
-    showYouTubePlayer(vid, sharedVideo, onYouTubeStateChange);
-    syncStatus.textContent = 'YouTube video sent to partner...';
-  } else {
-    hideYouTubePlayer(sharedVideo);
-    sharedVideo.src = url;
-    syncStatus.textContent = 'Video sent to partner...';
-  }
-  // Notify partner
-  if (roomId) {
-    db.ref(`rooms/${roomId}/stream`).set({ url });
-  }
-
+  loadStreamSync({ roomId, url, sharedVideo, syncStatus });
   saveCurrentState();
-}
-
-// --- YOUTUBE SYNC ---
-function onYouTubeStateChange(event) {
-  const ytPlayer = getYTPlayer();
-  const ytReady = getYTReady();
-  if (!roomId || !ytReady || !ytPlayer) return;
-  // 1=play, 2=pause
-  if (event.data === YT.PlayerState.PLAYING) {
-    db.ref(`rooms/${roomId}/stream`).update({
-      playing: true,
-      time: ytPlayer.getCurrentTime(),
-    });
-  } else if (event.data === YT.PlayerState.PAUSED) {
-    db.ref(`rooms/${roomId}/stream`).update({
-      playing: false,
-      time: ytPlayer.getCurrentTime(),
-    });
-  }
-}
-
-function acceptSharedVideo() {
-  const url = streamUrlInput.value;
-  if (isYouTubeUrl(url)) {
-    const vid = getYouTubeId(url);
-    showYouTubePlayer(vid, sharedVideo, onYouTubeStateChange);
-    syncStatus.textContent = 'Loading YouTube video...';
-  } else {
-    hideYouTubePlayer(sharedVideo);
-    sharedVideo.src = url;
-    syncStatus.textContent = 'Loading shared video...';
-  }
-  syncStatus.style.background = '#2a2a2a';
-}
-
-function setupWatchSync() {
-  if (!roomId) return;
-  const roomRef = db.ref(`rooms/${roomId}`);
-
-  // Listen for stream URL changes
-  roomRef.child('stream/url').on('value', (snapshot) => {
-    const url = snapshot.val();
-    if (url && url !== streamUrlInput.value) {
-      streamUrlInput.value = url;
-      syncStatus.innerHTML = `Partner shared a video: <button id="accept-shared-video" style="margin-left: 10px; padding: 5px 15px;">✓ Accept & Load</button>`;
-      const btn = document.getElementById('accept-shared-video');
-      btn.addEventListener('click', acceptSharedVideo);
-      syncStatus.style.background = '#2196f3';
-    }
-  });
-
-  // Listen for play/pause
-  roomRef.child('stream/playing').on('value', async (snapshot) => {
-    if (isSyncing) return;
-    const playing = snapshot.val();
-    const url = streamUrlInput.value;
-    const ytPlayer = getYTPlayer();
-    const ytReady = getYTReady();
-    if (isYouTubeUrl(url) && ytPlayer && ytReady) {
-      if (
-        playing === true &&
-        ytPlayer.getPlayerState() !== YT.PlayerState.PLAYING
-      ) {
-        ytPlayer.playVideo();
-        syncStatus.textContent = 'Playing in sync';
-      } else if (
-        playing === false &&
-        ytPlayer.getPlayerState() === YT.PlayerState.PLAYING
-      ) {
-        ytPlayer.pauseVideo();
-        syncStatus.textContent = 'Partner pressed pause';
-      }
-    } else {
-      if (playing === true && sharedVideo.paused) {
-        try {
-          await sharedVideo.play();
-          syncStatus.textContent = 'Playing in sync';
-        } catch (error) {
-          syncStatus.textContent = '▶️ Tap the video to start playing';
-          syncStatus.style.background = '#FF5722';
-          syncStatus.style.fontSize = '16px';
-          const clearWarning = () => {
-            syncStatus.style.background = '#2a2a2a';
-            syncStatus.style.fontSize = '14px';
-            sharedVideo.removeEventListener('play', clearWarning);
-          };
-          sharedVideo.addEventListener('play', clearWarning);
-        }
-      } else if (playing === false && !sharedVideo.paused) {
-        sharedVideo.pause();
-        syncStatus.textContent = 'Partner pressed pause';
-      }
-    }
-  });
-
-  // Listen for seek
-  roomRef.child('stream/time').on('value', (snapshot) => {
-    if (isSyncing) return;
-    const time = snapshot.val();
-    const url = streamUrlInput.value;
-    const ytPlayer = getYTPlayer();
-    const ytReady = getYTReady();
-    if (isYouTubeUrl(url) && ytPlayer && ytReady) {
-      if (time !== null && Math.abs(ytPlayer.getCurrentTime() - time) > 2) {
-        ytPlayer.seekTo(time, true);
-        syncStatus.textContent = `Syncing to ${Math.floor(time)}s`;
-      }
-    } else {
-      if (time !== null && Math.abs(sharedVideo.currentTime - time) > 2) {
-        sharedVideo.currentTime = time;
-        syncStatus.textContent = `Syncing to ${Math.floor(time)}s`;
-      }
-    }
-  });
-
-  // Send local events to Firebase
-  sharedVideo.addEventListener('play', () => {
-    if (roomId && !isSyncing) {
-      isSyncing = true;
-      db.ref(`rooms/${roomId}/stream`).update({
-        playing: true,
-        time: sharedVideo.currentTime,
-      });
-      setTimeout(() => (isSyncing = false), 1000);
-    }
-  });
-  sharedVideo.addEventListener('pause', () => {
-    if (roomId && !isSyncing) {
-      isSyncing = true;
-      db.ref(`rooms/${roomId}/stream`).update({
-        playing: false,
-        time: sharedVideo.currentTime,
-      });
-      setTimeout(() => (isSyncing = false), 1000);
-    }
-  });
-  sharedVideo.addEventListener('seeked', () => {
-    if (roomId && !isSyncing) {
-      isSyncing = true;
-      db.ref(`rooms/${roomId}/stream`).update({
-        time: sharedVideo.currentTime,
-      });
-      setTimeout(() => (isSyncing = false), 1000);
-    }
-  });
-  sharedVideo.addEventListener('loadeddata', () => {
-    syncStatus.textContent = 'Video loaded! Press play to start.';
-  });
-  sharedVideo.addEventListener('waiting', () => {
-    syncStatus.textContent = 'Buffering...';
-  });
-  sharedVideo.addEventListener('playing', () => {
-    syncStatus.textContent = 'Playing in sync';
-  });
 }
 
 // Add event handlers for fullscreen and mute buttons
