+++ src/lib/watch-sync.js
@@ -1,0 +1,250 @@
+// watch-sync.js - Watch mode and video sync feature
+
+import { db } from './firebase.js';
+
+import {
+  isYouTubeUrl,
+  getYouTubeId,
+  showYouTubePlayer,
+  hideYouTubePlayer,
+  getYTPlayer,
+  getYTReady,
+} from './video/youtube.js';
+
+// ===== STATE =====
+const state = {
+  watchMode: false,
+  isSyncing: false,
+};
+
+// ===== PUBLIC API =====
+
+export function getWatchMode() {
+  return state.watchMode;
+}
+
+export function getIsSyncing() {
+  return state.isSyncing;
+}
+
+export function toggleWatchMode({
+  videoContainer,
+  watchContainer,
+  toggleModeBtn,
+  sharedVideo,
+  streamUrlInput,
+  syncStatus,
+}) {
+  state.watchMode = !state.watchMode;
+
+  if (state.watchMode) {
+    videoContainer.style.display = 'none';
+    watchContainer.style.display = 'block';
+    toggleModeBtn.textContent = 'Switch to Video Chat';
+    syncStatus.textContent = 'Paste the same stream URL as your partner';
+  } else {
+    videoContainer.style.display = 'flex';
+    watchContainer.style.display = 'none';
+    toggleModeBtn.textContent = 'Switch to Watch Mode';
+    sharedVideo.src = '';
+    streamUrlInput.value = '';
+  }
+
+  return state.watchMode;
+}
+
+export function loadStream({ roomId, url, sharedVideo, syncStatus }) {
+  if (!url) {
+    syncStatus.textContent = 'Please enter a stream URL';
+    return;
+  }
+
+  if (isYouTubeUrl(url)) {
+    const vid = getYouTubeId(url);
+    showYouTubePlayer(vid, sharedVideo, (event) =>
+      onYouTubeStateChange(event, roomId)
+    );
+    syncStatus.textContent = 'YouTube video sent to partner...';
+  } else {
+    hideYouTubePlayer(sharedVideo);
+    sharedVideo.src = url;
+    syncStatus.textContent = 'Video sent to partner...';
+  }
+
+  if (roomId) {
+    db.ref(`rooms/${roomId}/stream`).set({ url });
+  }
+}
+
+export function setupWatchSync({
+  roomId,
+  sharedVideo,
+  streamUrlInput,
+  syncStatus,
+}) {
+  if (!roomId) return;
+
+  const roomRef = db.ref(`rooms/${roomId}`);
+
+  // Listen for stream URL changes
+  roomRef.child('stream/url').on('value', (snapshot) => {
+    const url = snapshot.val();
+    if (url && url !== streamUrlInput.value) {
+      streamUrlInput.value = url;
+      syncStatus.innerHTML = `Partner shared a video: <button id="accept-shared-video" style="margin-left: 10px; padding: 5px 15px;">✓ Accept & Load</button>`;
+      const btn = document.getElementById('accept-shared-video');
+      btn.addEventListener('click', () =>
+        acceptSharedVideo({ url, sharedVideo, syncStatus })
+      );
+      syncStatus.style.background = '#2196f3';
+    }
+  });
+
+  // Listen for play/pause
+  roomRef.child('stream/playing').on('value', async (snapshot) => {
+    if (state.isSyncing) return;
+    const playing = snapshot.val();
+    const url = streamUrlInput.value;
+    const ytPlayer = getYTPlayer();
+    const ytReady = getYTReady();
+
+    if (isYouTubeUrl(url) && ytPlayer && ytReady) {
+      if (
+        playing === true &&
+        ytPlayer.getPlayerState() !== YT.PlayerState.PLAYING
+      ) {
+        ytPlayer.playVideo();
+        syncStatus.textContent = 'Playing in sync';
+      } else if (
+        playing === false &&
+        ytPlayer.getPlayerState() === YT.PlayerState.PLAYING
+      ) {
+        ytPlayer.pauseVideo();
+        syncStatus.textContent = 'Partner pressed pause';
+      }
+    } else {
+      if (playing === true && sharedVideo.paused) {
+        try {
+          await sharedVideo.play();
+          syncStatus.textContent = 'Playing in sync';
+        } catch (error) {
+          syncStatus.textContent = '▶️ Tap the video to start playing';
+          syncStatus.style.background = '#FF5722';
+          syncStatus.style.fontSize = '16px';
+          const clearWarning = () => {
+            syncStatus.style.background = '#2a2a2a';
+            syncStatus.style.fontSize = '14px';
+            sharedVideo.removeEventListener('play', clearWarning);
+          };
+          sharedVideo.addEventListener('play', clearWarning);
+        }
+      } else if (playing === false && !sharedVideo.paused) {
+        sharedVideo.pause();
+        syncStatus.textContent = 'Partner pressed pause';
+      }
+    }
+  });
+
+  // Listen for seek
+  roomRef.child('stream/time').on('value', (snapshot) => {
+    if (state.isSyncing) return;
+    const time = snapshot.val();
+    const url = streamUrlInput.value;
+    const ytPlayer = getYTPlayer();
+    const ytReady = getYTReady();
+
+    if (isYouTubeUrl(url) && ytPlayer && ytReady) {
+      if (time !== null && Math.abs(ytPlayer.getCurrentTime() - time) > 2) {
+        ytPlayer.seekTo(time, true);
+        syncStatus.textContent = `Syncing to ${Math.floor(time)}s`;
+      }
+    } else {
+      if (time !== null && Math.abs(sharedVideo.currentTime - time) > 2) {
+        sharedVideo.currentTime = time;
+        syncStatus.textContent = `Syncing to ${Math.floor(time)}s`;
+      }
+    }
+  });
+
+  // Send local events to Firebase
+  setupLocalVideoListeners({ roomId, sharedVideo });
+}
+
+// ===== PRIVATE HELPERS =====
+
+function onYouTubeStateChange(event, roomId) {
+  const ytPlayer = getYTPlayer();
+  const ytReady = getYTReady();
+  if (!roomId || !ytReady || !ytPlayer) return;
+
+  if (event.data === YT.PlayerState.PLAYING) {
+    db.ref(`rooms/${roomId}/stream`).update({
+      playing: true,
+      time: ytPlayer.getCurrentTime(),
+    });
+  } else if (event.data === YT.PlayerState.PAUSED) {
+    db.ref(`rooms/${roomId}/stream`).update({
+      playing: false,
+      time: ytPlayer.getCurrentTime(),
+    });
+  }
+}
+
+function acceptSharedVideo({ url, sharedVideo, syncStatus }) {
+  if (isYouTubeUrl(url)) {
+    const vid = getYouTubeId(url);
+    showYouTubePlayer(vid, sharedVideo, onYouTubeStateChange);
+    syncStatus.textContent = 'Loading YouTube video...';
+  } else {
+    hideYouTubePlayer(sharedVideo);
+    sharedVideo.src = url;
+    syncStatus.textContent = 'Loading shared video...';
+  }
+  syncStatus.style.background = '#2a2a2a';
+}
+
+function setupLocalVideoListeners({ roomId, sharedVideo }) {
+  sharedVideo.addEventListener('play', () => {
+    if (roomId && !state.isSyncing) {
+      state.isSyncing = true;
+      db.ref(`rooms/${roomId}/stream`).update({
+        playing: true,
+        time: sharedVideo.currentTime,
+      });
+      setTimeout(() => (state.isSyncing = false), 1000);
+    }
+  });
+
+  sharedVideo.addEventListener('pause', () => {
+    if (roomId && !state.isSyncing) {
+      state.isSyncing = true;
+      db.ref(`rooms/${roomId}/stream`).update({
+        playing: false,
+        time: sharedVideo.currentTime,
+      });
+      setTimeout(() => (state.isSyncing = false), 1000);
+    }
+  });
+
+  sharedVideo.addEventListener('seeked', () => {
+    if (roomId && !state.isSyncing) {
+      state.isSyncing = true;
+      db.ref(`rooms/${roomId}/stream`).update({
+        time: sharedVideo.currentTime,
+      });
+      setTimeout(() => (state.isSyncing = false), 1000);
+    }
+  });
+
+  sharedVideo.addEventListener('loadeddata', () => {
+    // Could pass syncStatus here if needed
+  });
+
+  sharedVideo.addEventListener('waiting', () => {
+    // Could pass syncStatus here if needed
+  });
+
+  sharedVideo.addEventListener('playing', () => {
+    // Could pass syncStatus here if needed
+  });
+}
